# contracts/openapi/core-api.v1.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "kinsun.ai Core API",
    "version": "1.2.0",
    "summary": "Implemented Core Domain, consent, security and outbox APIs.",
    "description": "Current executable Core API contract. Every protected operation re-evaluates tenant, elder, relationship or assignment scope. Cognito access tokens authenticate existing actors; the onboarding resolver separately accepts a verified Cognito ID token and never treats persona intent as authorization."
  },
  "paths": {
...
    "/api/v1/elders/{elder_id}/family-invitations": {
      "get": {
        "tags": ["family-invitations"],
        "summary": "List safe family invitation statuses",
        "description": "Elder-only, live tenant-scoped listing. The plaintext invitation code, token hash, email digest and redeemer identity are never returned.",
        "operationId": "list_family_invitations_api_v1_elders__elder_id__family_invitations_get",
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          }
        ],
        "responses": {
          "200": {
            "description": "Authorized invitation status list.",
            "content": {
              "application/json": {
                "schema": {"$ref": "../schemas/common/FamilyInvitationListEnvelopeV1.json"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFoundOrForbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/ValidationFailed"},
          "503": {"$ref": "#/components/responses/Unavailable"}
        },
        "security": [{"bearerAuth": []}]
      },
      "post": {
        "tags": ["family-invitations"],
        "summary": "Create one-time family invitation",
        "description": "Elder-only and consent-gated. Returns the plaintext code once; Core persists only its keyed digest.",
        "operationId": "create_family_invitation_api_v1_elders__elder_id__family_invitations_post",
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {"type": "string", "minLength": 1, "maxLength": 160}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "../schemas/domain/CreateFamilyInvitationRequestV1.json"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitation issued; copy the one-time code now.",
            "content": {
              "application/json": {
                "schema": {"$ref": "../schemas/common/FamilyInvitationCreatedEnvelopeV1.json"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFoundOrForbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/ValidationFailed"},
          "503": {"$ref": "#/components/responses/Unavailable"}
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/api/v1/elders/{elder_id}/family-invitations/{invitation_id}/revoke": {
      "post": {
        "tags": ["family-invitations"],
        "summary": "Revoke an issued family invitation",
        "description": "Elder-only, idempotent revocation. A redeemed, expired or locked invitation cannot be converted to revoked.",
        "operationId": "revoke_family_invitation_api_v1_elders__elder_id__family_invitations__invitation_id__revoke_post",
        "parameters": [
          {
            "name": "elder_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          },
          {
            "name": "invitation_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string", "format": "uuid"}
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {"type": "string", "minLength": 1, "maxLength": 160}
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation is revoked, including a safe replay of the prior revocation.",
            "content": {
              "application/json": {
                "schema": {"$ref": "../schemas/common/FamilyInvitationStatusEnvelopeV1.json"}
              }
            }
          },
          "401": {"$ref": "#/components/responses/Unauthorized"},
          "403": {"$ref": "#/components/responses/Forbidden"},
          "404": {"$ref": "#/components/responses/NotFoundOrForbidden"},
          "409": {"$ref": "#/components/responses/Conflict"},
          "422": {"$ref": "#/components/responses/ValidationFailed"},
          "503": {"$ref": "#/components/responses/Unavailable"}
        },
        "security": [{"bearerAuth": []}]
      }
    },
    "/api/v1/me": {
      "get": {
        "tags": [
          "identity"
        ],
        "summary": "Get Me",
        "description": "Return the authenticated actor's profile.\n\nIncludes actor identity fields and the list of care unit IDs\nthe actor belongs to.",
        "operationId": "get_me_api_v1_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/ActorProfileEnvelopeV1.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundOrForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/v1/me/authorized-elders": {
      "get": {
        "tags": [
          "identity"
        ],
        "summary": "Get Authorized Elders",
        "description": "Return paginated elders the actor is authorized to access.\n\nMode must be compatible with the actor's role. Incompatible\ncombinations return 403 via RoleModeIncompatibleError.",
        "operationId": "get_authorized_elders_api_v1_me_authorized_elders_get",
        "parameters": [
          {
            "name": "mode",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ElderMode",
              "description": "Access mode: daycare, home-care, or family"
            },
            "description": "Access mode: daycare, home-care, or family"
          },
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination cursor",
              "title": "Cursor"
            },
            "description": "Opaque pagination cursor"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Page size (1–100)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Page size (1–100)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "../schemas/common/AuthorizedElderListEnvelopeV1.json"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFoundOrForbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/ValidationFailed"
          },
          "503": {
            "$ref": "#/components/responses/Unavailable"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ElderMode": {
        "type": "string",
        "enum": [
          "daycare",
          "home-care",
          "family"
        ],
        "title": "ElderMode",
        "description": "Valid mode values for the authorized-elders query parameter."
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito access token. Core verifies signature, issuer, expiry, token_use=access and client_id, then resolves actor, tenant and role from live Core database state. Development fake auth requires the explicit FAKE_AUTH_ENABLED flag."
      },
      "cognitoIdToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Cognito ID token accepted only by onboarding resolution. Core verifies signature, issuer, expiry, token_use=id, audience and verified email; claims do not directly grant any role or elder scope."
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Authentication required or authenticator unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "Forbidden": {
        "description": "The actor is inactive or its role cannot perform the operation.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "NotFoundOrForbidden": {
        "description": "Resource missing or outside live authorization scope.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "Conflict": {
        "description": "State, version, or idempotency conflict.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "ValidationFailed": {
        "description": "Schema or semantic validation failed.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      },
      "Unavailable": {
        "description": "Required dependency unavailable.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "../schemas/common/ErrorEnvelopeV1.json"
            }
          }
        }
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}